Object-Oriented-Programming-with-Java-pt1-University-of Helsinki-moocfi/week2-038.ManyPrints/src/ManyPrints.java

import java.util.Scanner;

public class ManyPrints {
    // NOTE: do not change the method definition, e.g. add parameters to method
    public static void printText() {
        System.out.println("In the beginning there were the swamp, the hoe and Java.");
    }

    public static void main(String[] args) {
        // ask the user how many times the text should be printed
        // use the while structure to call the printText method several times
        Scanner reader = new Scanner(System.in);
        System.out.print("How many?\n");
        int num = Integer.parseInt(reader.nextLine());
        for(int i = 0; i<num; i++){printText();}
    }
}
